Implemented Hebrew calendar
Hebrew calendar now works after long, arduous process of iteration, bug-hunting, and accounting for fencepost errors. I wasn't even able to find every error, so in the end I had to kludge in a test in the event of a negative year to check if the Julian Day derived corresponds to 1 Tishrei in the current year, and add 1 if it doesn't. The comment calls this some "sneaky buggery".