Replaced hebrew.py with jewish.py
Wrote a brand new implementation of the Jewish Hebrew calendar algorithms. This one is named jewish.py rather than hebrew.py because the Jewish calendar isn't the only Hebrew calendar.
The new algorithm is simpler and more straightforward. There also seems to have been an obscure bug in hebrew.py which jewish.py doesn't seem to have. Plus, there are no random fencepost errors that apparently depend on the actual phase of the moon, which means this algorithm is free of kludges.