Refactored Samaritan algorithm
Completely rewrote samaritan.py, using the methods from jewish.py; in fact, the new samaritan.py is jewish.py with a different epoch and lacking the dechiyyot. The new algorithm is cleaner and more elegant, and without need for kludges to account for randomly-gained days. It sometimes differs by a day from the old algorithm, and years may sometimes have an extra day depending on when the molad of Nisan falls; I believe this is closer to the system used by the Samaritans.