Added proper motion accounting to sunmoon.f90
Added a subroutine to sunmoon.f90 to account for stars' proper motion in space. This appears to be more accurate than simply calculating based on precession of the equinoxes. This will be of use in calculating sidereal times, with the caveat that over longer time periods, it becomes less and less accurate (but then we do live in a chaotic universe)