Bugfix
The years 1601—1621 were completely out of bounds.  It's possible I was supposed to insert something for those specific years, but  I'm buggered if I can think of what it is, and since I'm an antisocial weirdo who can't persuade anybody to do fun things to my bum, I just kludged it.