Rewrote Thandeikta
Completely rewrote thandeikta.py from scratch. It seem to give correct dates now.

Rather than using the formal rules to determine the length of the year, this programme uses an algorithm that checks the length of the specified year and the adjacent years. If one of them is outside the permitted year lengths, it moves the Julian Day of the 1st of Tagu forward or back one day to ensure proper lengths.