Added ancient Greek calendars
Added various calendars of ancient Greece. These are all kata theon (per the gods), as opposed to kata archon (per the magistrates). The calendars of the gods were strictly observational, while the calendars of the magistrates had days and months added in accordance with political and logistical expedience. All these calendars repeat the middle month in leap years; while this was the most common approach in ancient Greece, the leap month could be inserted at any point in the year.

Also I properly added the Hermetic Lunar Week calendar to calconv.py; I had forgotten to search and replace the placeholder variable names :~